Fan service for SFMono-aficionados
SF Mono is not a system font. It's available in macOS, but it can't be used
outside of Terminal.app. However, you can install it as a system font by
importing the font files from the Terminal.app application bundle.

    $ open /Applications/Utilities/Terminal.app/Contents/Resources/Fonts/*

A Font Book window will pop up, allowing you to install the font. The validation
window will tell you not to install this font because an error has been found
for each variant, but you can install it anyway.

After completing the steps above, you're officially promoted to an
SFMono-aficionado, and enough.css spends 15 bytes on treating you to your
favorite monospace font.