Update link colors, remove color-scheme, color <hr>
Instead of setting `color-scheme` in the stylesheet, enabling dark mode
should be done on the page where enough.dark.css is included through a
`<meta>` tag:

    <meta name="color-scheme" content="light dark">

Instead of adding darkmode styling to a webpage, enough.dak.css's goal
is to provide a basic, legible stylesheet for dark mode, like the
default browser style sheet is for pages in "light" mode.

To do this, it updates the colors for links and already-visited links to
the default stylesheets closest color that has a contrast rating of 7.0
on a black background.

It also styles `<hr>` tags to use darker grays that work on dark
backgrounds.