Keep code tags the same font size, reduce pre block font size (#7)
* Keep code tags the same font size, reduce pre block font size

Instead of reducing the font size for all code to 0.85 em, use regular
size for code tags, but reduce preformatted block font sizes to 0.8 to
/almost/ fit 80 characters in code blocks.

* Move font setup to body instead of html

By keeping code tag font sizes 1 em outside of preformatted blocks,
enough.css no longer needs to use the rem unit. Because of that, the
font setup can be moved back to the body tag.