Use ems for monospace elements
Previously, enough.css used the rem unit to decrease the size of
monospace fonts. It chose rems because that prevented code elements
inside of pre blocks to be rendered smaller than elements without it.
However, that approach also caused code elements in headings to be
rendered at 0.8 rems, which is strange for a 2.0 rem headline.

Instead of the previous approach, use a 0.8 em font size for all code
blocks, but add an exception for code elements nested within pre
blocks.